node新增messge字段, 代码结构拆分
