+++ src/agentscope/web/workstation/workflow_node.py
@@ -155,6 +155,7 @@ class ASDiGraph(nx.DiGraph):
             node_result_dict = {
                 "node_id": node["id"],
                 "node_status": "",
+                "node_message": "",
                 "node_type": node["type"],
                 "inputs": {},
                 "outputs": {},
@@ -164,17 +165,18 @@ class ASDiGraph(nx.DiGraph):
         return node_result
 
     @staticmethod
-    def update_nodes_with_values(nodes, output_values, input_values, status_values):
+    def update_nodes_with_values(nodes, output_values, input_values, status_values, message_values):
         for index, node in enumerate(nodes):
             node_id = node['node_id']
             node['node_status'] = status_values.get(node_id, "init")
+            node['node_message'] = message_values.get(node_id, "")
             node['inputs'] = input_values.get(node_id, {})
             node['outputs'] = output_values.get(node_id, {})
             nodes[index] = node
 
         return nodes
 
-    def save(self, save_filepath: str = "",) -> None:
+    def save(self, save_filepath: str = "", ) -> None:
         if len(self.config) > 0:
             # Write the script to file
             with open(save_filepath, "w", encoding="utf-8") as file:
@@ -234,7 +236,7 @@ class ASDiGraph(nx.DiGraph):
         ]
         logger.info(f"topological sorted nodes: {sorted_nodes}")
 
-        total_input_values, total_output_values, total_status_values = {}, {}, {}
+        total_input_values, total_output_values, total_status_values, total_message_values = {}, {}, {}, {}
         output_values = {}
         # Run with predecessors outputs
         for node_id in sorted_nodes:
@@ -261,18 +263,19 @@ class ASDiGraph(nx.DiGraph):
             total_input_values[node_id] = self.nodes[node_id]["opt"].input_params
             total_output_values[node_id] = self.nodes[node_id]["opt"].output_params
             total_status_values[node_id] = self.nodes[node_id]["opt"].running_status
+            total_message_values[node_id] = self.nodes[node_id]["opt"].running_message
 
         # 初始化节点运行结果并更新
         nodes_result = ASDiGraph.set_initial_nodes_result(config)
         updated_nodes_result = ASDiGraph.update_nodes_with_values(
-            nodes_result, total_output_values, total_input_values, total_status_values)
+            nodes_result, total_output_values, total_input_values, total_status_values, total_message_values)
         logger.info(f"workflow total running result: {updated_nodes_result}")
         return total_input_values[sorted_nodes[-1]], updated_nodes_result
 
     def compile(  # type: ignore[no-untyped-def]
-        self,
-        compiled_filename: str = "",
-        **kwargs,
+            self,
+            compiled_filename: str = "",
+            **kwargs,
     ) -> str:
         """Compile DAG to a runnable python code"""
 
@@ -326,10 +329,10 @@ class ASDiGraph(nx.DiGraph):
 
     # pylint: disable=R0912
     def add_as_node(
-        self,
-        node_id: str,
-        node_info: dict,
-        config: dict,
+            self,
+            node_id: str,
+            node_info: dict,
+            config: dict,
     ) -> Any:
         """
         Add a node to the graph based on provided node information and
@@ -452,11 +455,11 @@ class WorkflowNode(ABC):
     node_type = None
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         """
         Initialize nodes. Implement specific initialization logic in
@@ -499,11 +502,11 @@ class ModelNode(WorkflowNode):
     node_type = WorkflowNodeType.MODEL
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         read_model_configs([self.opt_kwargs])
@@ -527,11 +530,11 @@ class MsgNode(WorkflowNode):
     node_type = WorkflowNodeType.MESSAGE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.msg = Msg(**self.opt_kwargs)
@@ -543,7 +546,7 @@ class MsgNode(WorkflowNode):
         return {
             "imports": "from agentscope.message import Msg",
             "inits": f"{DEFAULT_FLOW_VAR} = Msg"
-            f"({kwarg_converter(self.opt_kwargs)})",
+                     f"({kwarg_converter(self.opt_kwargs)})",
             "execs": "",
         }
 
@@ -556,11 +559,11 @@ class DialogAgentNode(WorkflowNode):
     node_type = WorkflowNodeType.AGENT
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = DialogAgent(**self.opt_kwargs)
@@ -572,9 +575,9 @@ class DialogAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import DialogAgent",
             "inits": f"{self.var_name} = DialogAgent("
-            f"{kwarg_converter(self.opt_kwargs)})",
+                     f"{kwarg_converter(self.opt_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -586,11 +589,11 @@ class UserAgentNode(WorkflowNode):
     node_type = WorkflowNodeType.AGENT
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = UserAgent(**self.opt_kwargs)
@@ -602,9 +605,9 @@ class UserAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import UserAgent",
             "inits": f"{self.var_name} = UserAgent("
-            f"{kwarg_converter(self.opt_kwargs)})",
+                     f"{kwarg_converter(self.opt_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -616,11 +619,11 @@ class TextToImageAgentNode(WorkflowNode):
     node_type = WorkflowNodeType.AGENT
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = TextToImageAgent(**self.opt_kwargs)
@@ -632,9 +635,9 @@ class TextToImageAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import TextToImageAgent",
             "inits": f"{self.var_name} = TextToImageAgent("
-            f"{kwarg_converter(self.opt_kwargs)})",
+                     f"{kwarg_converter(self.opt_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -646,11 +649,11 @@ class DictDialogAgentNode(WorkflowNode):
     node_type = WorkflowNodeType.AGENT
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = DictDialogAgent(**self.opt_kwargs)
@@ -662,9 +665,9 @@ class DictDialogAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import DictDialogAgent",
             "inits": f"{self.var_name} = DictDialogAgent("
-            f"{kwarg_converter(self.opt_kwargs)})",
+                     f"{kwarg_converter(self.opt_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -676,11 +679,11 @@ class ReActAgentNode(WorkflowNode):
     node_type = WorkflowNodeType.AGENT
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         # Build tools
@@ -698,10 +701,10 @@ class ReActAgentNode(WorkflowNode):
         return {
             "imports": "from agentscope.agents import ReActAgent",
             "inits": f"{self.var_name} = ReActAgent"
-            f"({kwarg_converter(self.opt_kwargs)}, tools"
-            f"={deps_converter(self.dep_vars)})",
+                     f"({kwarg_converter(self.opt_kwargs)}, tools"
+                     f"={deps_converter(self.dep_vars)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -716,11 +719,11 @@ class MsgHubNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.announcement = Msg(
@@ -759,7 +762,7 @@ class MsgHubNode(WorkflowNode):
         """
         return {
             "imports": "from agentscope.msghub import msghub\n"
-            "from agentscope.message import Msg",
+                       "from agentscope.message import Msg",
             "inits": "",
             "execs": execs,
         }
@@ -776,11 +779,11 @@ class PlaceHolderNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = placeholder
@@ -791,10 +794,10 @@ class PlaceHolderNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.pipelines.functional import "
-            "placeholder",
+                       "placeholder",
             "inits": f"{self.var_name} = placeholder",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -809,11 +812,11 @@ class SequentialPipelineNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.pipeline = SequentialPipeline(operators=self.dep_opts)
@@ -825,9 +828,9 @@ class SequentialPipelineNode(WorkflowNode):
         return {
             "imports": "from agentscope.pipelines import SequentialPipeline",
             "inits": f"{self.var_name} = SequentialPipeline("
-            f"{deps_converter(self.dep_vars)})",
+                     f"{deps_converter(self.dep_vars)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -842,15 +845,15 @@ class ForLoopPipelineNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         assert (
-            len(self.dep_opts) == 1
+                len(self.dep_opts) == 1
         ), "ForLoopPipelineNode can only contain one PipelineNode."
         self.pipeline = ForLoopPipeline(
             loop_body_operators=self.dep_opts[0],
@@ -864,11 +867,11 @@ class ForLoopPipelineNode(WorkflowNode):
         return {
             "imports": "from agentscope.pipelines import ForLoopPipeline",
             "inits": f"{self.var_name} = ForLoopPipeline("
-            f"loop_body_operators="
-            f"{deps_converter(self.dep_vars)},"
-            f" {kwarg_converter(self.source_kwargs)})",
+                     f"loop_body_operators="
+                     f"{deps_converter(self.dep_vars)},"
+                     f" {kwarg_converter(self.source_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -883,15 +886,15 @@ class WhileLoopPipelineNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         assert (
-            len(self.dep_opts) == 1
+                len(self.dep_opts) == 1
         ), "WhileLoopPipelineNode can only contain one PipelineNode."
         self.pipeline = WhileLoopPipeline(
             loop_body_operators=self.dep_opts[0],
@@ -905,11 +908,11 @@ class WhileLoopPipelineNode(WorkflowNode):
         return {
             "imports": "from agentscope.pipelines import WhileLoopPipeline",
             "inits": f"{self.var_name} = WhileLoopPipeline("
-            f"loop_body_operators="
-            f"{deps_converter(self.dep_vars)},"
-            f" {kwarg_converter(self.source_kwargs)})",
+                     f"loop_body_operators="
+                     f"{deps_converter(self.dep_vars)},"
+                     f" {kwarg_converter(self.source_kwargs)})",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.var_name}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -924,15 +927,15 @@ class IfElsePipelineNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         assert (
-            0 < len(self.dep_opts) <= 2
+                0 < len(self.dep_opts) <= 2
         ), "IfElsePipelineNode must contain one or two PipelineNode."
         if len(self.dep_opts) == 1:
             self.pipeline = IfElsePipeline(
@@ -956,15 +959,15 @@ class IfElsePipelineNode(WorkflowNode):
             return {
                 "imports": imports,
                 "inits": f"{self.var_name} = IfElsePipeline("
-                f"if_body_operators={self.dep_vars[0]})",
+                         f"if_body_operators={self.dep_vars[0]})",
                 "execs": execs,
             }
         elif len(self.dep_vars) == 2:
             return {
                 "imports": imports,
                 "inits": f"{self.var_name} = IfElsePipeline("
-                f"if_body_operators={self.dep_vars[0]}, "
-                f"else_body_operators={self.dep_vars[1]})",
+                         f"if_body_operators={self.dep_vars[0]}, "
+                         f"else_body_operators={self.dep_vars[1]})",
                 "execs": execs,
             }
         raise ValueError
@@ -981,11 +984,11 @@ class SwitchPipelineNode(WorkflowNode):
     node_type = WorkflowNodeType.PIPELINE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         assert 0 < len(self.dep_opts), (
@@ -1009,9 +1012,9 @@ class SwitchPipelineNode(WorkflowNode):
             )
 
         for key, value, var in zip(
-            self.opt_kwargs["cases"],
-            self.dep_opts,
-            self.dep_vars,
+                self.opt_kwargs["cases"],
+                self.dep_opts,
+                self.dep_vars,
         ):
             case_operators[key] = value.pipeline
             self.case_operators_var[key] = var
@@ -1035,9 +1038,9 @@ class SwitchPipelineNode(WorkflowNode):
         return {
             "imports": imports,
             "inits": f"{self.var_name} = SwitchPipeline(case_operators="
-            f"{dict_converter(self.case_operators_var)}, "
-            f"default_operators={self.default_var_name},"
-            f" {kwarg_converter(self.source_kwargs)})",
+                     f"{dict_converter(self.case_operators_var)}, "
+                     f"default_operators={self.default_var_name},"
+                     f" {kwarg_converter(self.source_kwargs)})",
             "execs": execs,
         }
 
@@ -1054,11 +1057,11 @@ class CopyNode(WorkflowNode):
     node_type = WorkflowNodeType.COPY
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         assert len(self.dep_opts) == 1, "CopyNode can only have one parent!"
@@ -1072,7 +1075,7 @@ class CopyNode(WorkflowNode):
             "imports": "",
             "inits": "",
             "execs": f"{DEFAULT_FLOW_VAR} = {self.dep_vars[0]}"
-            f"({DEFAULT_FLOW_VAR})",
+                     f"({DEFAULT_FLOW_VAR})",
         }
 
 
@@ -1084,11 +1087,11 @@ class BingSearchServiceNode(WorkflowNode):
     node_type = WorkflowNodeType.SERVICE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.service_func = ServiceFactory.get(bing_search, **self.opt_kwargs)
@@ -1096,9 +1099,9 @@ class BingSearchServiceNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.service import ServiceFactory\n"
-            "from agentscope.service import bing_search",
+                       "from agentscope.service import bing_search",
             "inits": f"{self.var_name} = ServiceFactory.get(bing_search,"
-            f" {kwarg_converter(self.opt_kwargs)})",
+                     f" {kwarg_converter(self.opt_kwargs)})",
             "execs": "",
         }
 
@@ -1111,11 +1114,11 @@ class GoogleSearchServiceNode(WorkflowNode):
     node_type = WorkflowNodeType.SERVICE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.service_func = ServiceFactory.get(
@@ -1126,9 +1129,9 @@ class GoogleSearchServiceNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.service import ServiceFactory\n"
-            "from agentscope.service import google_search",
+                       "from agentscope.service import google_search",
             "inits": f"{self.var_name} = ServiceFactory.get(google_search,"
-            f" {kwarg_converter(self.opt_kwargs)})",
+                     f" {kwarg_converter(self.opt_kwargs)})",
             "execs": "",
         }
 
@@ -1141,11 +1144,11 @@ class PythonServiceNode(WorkflowNode):
     node_type = WorkflowNodeType.SERVICE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.service_func = ServiceFactory.get(execute_python_code)
@@ -1153,9 +1156,9 @@ class PythonServiceNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.service import ServiceFactory\n"
-            "from agentscope.service import execute_python_code",
+                       "from agentscope.service import execute_python_code",
             "inits": f"{self.var_name} = ServiceFactory.get("
-            f"execute_python_code)",
+                     f"execute_python_code)",
             "execs": "",
         }
 
@@ -1168,11 +1171,11 @@ class ReadTextServiceNode(WorkflowNode):
     node_type = WorkflowNodeType.SERVICE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.service_func = ServiceFactory.get(read_text_file)
@@ -1180,7 +1183,7 @@ class ReadTextServiceNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.service import ServiceFactory\n"
-            "from agentscope.service import read_text_file",
+                       "from agentscope.service import read_text_file",
             "inits": f"{self.var_name} = ServiceFactory.get(read_text_file)",
             "execs": "",
         }
@@ -1194,11 +1197,11 @@ class WriteTextServiceNode(WorkflowNode):
     node_type = WorkflowNodeType.SERVICE
 
     def __init__(
-        self,
-        node_id: str,
-        opt_kwargs: dict,
-        source_kwargs: dict,
-        dep_opts: list,
+            self,
+            node_id: str,
+            opt_kwargs: dict,
+            source_kwargs: dict,
+            dep_opts: list,
     ) -> None:
         super().__init__(node_id, opt_kwargs, source_kwargs, dep_opts)
         self.service_func = ServiceFactory.get(write_text_file)
@@ -1206,7 +1209,7 @@ class WriteTextServiceNode(WorkflowNode):
     def compile(self) -> dict:
         return {
             "imports": "from agentscope.service import ServiceFactory\n"
-            "from agentscope.service import write_text_file",
+                       "from agentscope.service import write_text_file",
             "inits": f"{self.var_name} = ServiceFactory.get(write_text_file)",
             "execs": "",
         }
@@ -1251,8 +1254,9 @@ class StartNode(WorkflowNode):
         # opt_kwargs 包含用户定义的节点输入变量
         self.input_params = {}
         self.output_params = {}
-        # init --> running -> success/failed:xxx
+        # init --> running -> success/failed
         self.running_status = "init"
+        self.running_message = ""
         self.dag_obj = dag_obj
         self.dag_id = ""
         if self.dag_obj:
@@ -1268,7 +1272,8 @@ class StartNode(WorkflowNode):
             return self.output_params
         except Exception as err:
             logger.error(f'{traceback.format_exc()}')
-            self.running_status = f'failed: {repr(err)}'
+            self.running_status = 'failed'
+            self.running_message = f'{repr(err)}'
             return {}
 
     def run(self, *args, **kwargs):
@@ -1363,6 +1368,7 @@ class EndNode(WorkflowNode):
         self.output_params = {}
         # init --> running -> success/failed:xxx
         self.running_status = "init"
+        self.running_message = ""
         self.dag_obj = dag_obj
         self.dag_id = ""
         if self.dag_obj:
@@ -1382,7 +1388,8 @@ class EndNode(WorkflowNode):
             return {}
         except Exception as err:
             logger.error(f'{traceback.format_exc()}')
-            self.running_status = f'failed: {repr(err)}'
+            self.running_status = 'failed'
+            self.running_message = f'{repr(err)}'
             return {}
 
     def run(self, *args, **kwargs):
@@ -1471,6 +1478,7 @@ class PythonServiceUserTypingNode(WorkflowNode):
         self.output_params_spec = {}
         # init --> running -> success/failed:xxx
         self.running_status = "init"
+        self.running_message = ""
         self.dag_obj = dag_obj
         self.dag_id = ""
         if self.dag_obj:
@@ -1491,7 +1499,8 @@ class PythonServiceUserTypingNode(WorkflowNode):
             return self.output_params
         except Exception as err:
             logger.error(f'{traceback.format_exc()}')
-            self.running_status = f'failed: {repr(err)}'
+            self.running_status = 'failed'
+            self.running_message = f'{repr(err)}'
             return {}
 
     def run(self, *args, **kwargs):
@@ -1636,6 +1645,7 @@ class ApiNode(WorkflowNode):
         self.output_params_spec = {}
         # init --> running -> success/failed:xxx
         self.running_status = "init"
+        self.running_message = ''
         self.dag_obj = dag_obj
         self.dag_id = ""
         if self.dag_obj:
@@ -1731,7 +1741,8 @@ class ApiNode(WorkflowNode):
             return self.output_params
         except Exception as err:
             logger.error(f'{traceback.format_exc()}')
-            self.running_status = f'failed: {repr(err)}'
+            self.running_status = 'failed'
+            self.running_message = f'{repr(err)}'
             return {}
 
     def run(self, *args, **kwargs):
@@ -1834,9 +1845,9 @@ NODE_NAME_MAPPING = {
 
 
 def get_all_agents(
-    node: WorkflowNode,
-    seen_agents: Optional[set] = None,
-    return_var: bool = False,
+        node: WorkflowNode,
+        seen_agents: Optional[set] = None,
+        return_var: bool = False,
 ) -> List:
     """
     Retrieve all unique agent objects from a pipeline.
